fabriguard/backend/requirements.txt
# FabriGuard Backend Dependencies

# Web Framework
fastapi==0.109.0
uvicorn[standard]==0.27.0
python-multipart==0.0.6
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4

# Database
sqlalchemy==2.0.25
alembic==1.13.1
asyncpg==0.29.0
psycopg2-binary==2.9.9

# Time Series Database
influxdb-client==1.40.0

# Redis for caching and queues
redis==5.0.1
celery==5.3.6

# ML/Data Science
numpy==1.26.3
pandas==2.1.4
scikit-learn==1.4.0
scipy==1.12.0
joblib==1.3.2

# Deep Learning (optional, for advanced models)
# torch==2.1.2
# tensorflow==2.15.0

# Data Validation
pydantic==2.5.3
pydantic-settings==2.1.0

# HTTP Client
httpx==0.26.0
aiohttp==3.9.1

# Notifications
firebase-admin==6.4.0
twilio==8.12.0

# Utilities
python-dotenv==1.0.0
structlog==24.1.0
tenacity==8.2.3
apscheduler==3.10.4

# Testing
pytest==7.4.4
pytest-asyncio==0.23.3
pytest-cov==4.1.0
httpx==0.26.0
factory-boy==3.3.0

# Development
black==24.1.1
isort==5.13.2
mypy==1.8.0
ruff==0.1.14
